+++ src/lib.rs
@@ -11,6 +11,7 @@ pub mod fdlimit;
 pub mod filter;
 pub mod hosted;
 pub mod humanize;
+pub mod live;
 pub mod netstats;
 pub mod selector;
 pub mod service;